Changed a couple tests files/made a directory for our tests. Added courseId foreign key to sectionTable
