+++ scheduler/models.py
@@ -37,6 +37,10 @@ class SectionTable(models.Model):
     name = models.CharField(max_length=30)
     time = models.CharField(max_length=30, blank=True, null=True)
 
+    courseId = models.ForeignKey(
+        CourseTable,
+        on_delete=models.CASCADE,
+    )
 
 class LabTable(models.Model):
     # Fields
